# src/main/java/com/projetoceos/review/entity/Review.java
package com.projetoceos.review.entity;

import javax.persistence.Column;
import javax.persistence.Entity;
import javax.persistence.GeneratedValue;
import javax.persistence.GenerationType;
import javax.persistence.Id;

import lombok.Getter;
import lombok.Setter;

@Entity
@Getter
@Setter
public class Review {

    @Id
    @GeneratedValue(strategy = GenerationType.AUTO)
    private long id;

    @Column(nullable = false)
    private int rating;

    @Column(nullable = true)
    private String title;

    @Column(nullable = true)
    private String description;

}