# src/main/resources/swagger/swagger-review-api.yml
swagger: "2.0"
info:
  description: "Uma API para cadastro de avaliações."
  version: "1.0.0"
  title: "Swagger Review"
host: "localhost:8081"
basePath: "/api"
tags:
- name: "Review"
  description: "Nossas APIs de avaliação."
schemes:
- "http"
paths:
  "/reviews":
    post:
      tags:
      - "Review"
      summary: "Adicionar uma avaliação."
      operationId: "createReview"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Objeto avaliação para ser criado."
        required: true
        schema:
          $ref: "#/definitions/Review"
      responses:
        "201":
          description: "Created"
  "/reviews/{id}":
    get:
      tags:
      - "Review"
      summary: "Encontrar avaliação pelo Id."
      description: "Retorna uma avaliação peli Id."
      operationId: "getReviewById"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "ID da avaliação."
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Sucesso!"
        "404":
          description: "Review not found."
    put:
      tags:
      - "Review"
      summary: "Atualizar uma avaliação"
      operationId: "updateReview"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "ID da avaliação a ser atualizada."
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "body"
        description: "Objeto avaliação para ser criado."
        required: true
        schema:
          $ref: "#/definitions/Review"
      responses:
        "200":
          description: "Review updated"
        "404":
          description: "Review not found"
    delete:
      tags:
      - "Review"
      summary: "Deletar uma avaliação."
      operationId: "deleteReview"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "Id da avaliação para ser deletada"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Deleted with success"
        "404":
          description: "Review not found"
definitions:
  Review:
    type: "object"
    required:
    - "rating"
    - "title"
    - "description"
    properties:
      id:
        type: "integer"
        format: "int64"
      title:
        type: "string"
        example: "Um título de avaliação."
      description:
        type: "string"
        example: "Uma descrição de avaliação."
      rating:
        type: "integer"
        description: "Uma nota de avaliação."
  ApiResponse:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      type:
        type: "string"
      message:
        type: "string"